requirements.txt
whisper
gtts
elevenlabs
pyaudio
ffmpeg-python
opencv-python
requests
gradio
openai
python-dotenv